[Refactor] : LearningProgress 코드 수정 (#69)
