+++ src/main/java/com/potato/balbambalbam/user/join/jwt/CustomAuthenticationProvider.java
@@ -2,7 +2,7 @@ package com.potato.balbambalbam.user.join.jwt;
 
 import com.potato.balbambalbam.data.entity.User;
 import com.potato.balbambalbam.data.repository.UserRepository;
-import com.potato.balbambalbam.user.join.jwt.dto.CustomUserDetails;
+import com.potato.balbambalbam.user.join.dto.CustomUserDetails;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
