+++ src/main/java/com/potato/balbambalbam/user/join/service/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package com.potato.balbambalbam.user.join.service;
 
 import com.potato.balbambalbam.data.entity.User;
 import com.potato.balbambalbam.data.repository.UserRepository;
-import com.potato.balbambalbam.user.join.jwt.dto.CustomUserDetails;
+import com.potato.balbambalbam.user.join.dto.CustomUserDetails;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
