+++ src/main/java/com/potato/balbambalbam/user/join/service/JoinService.java
@@ -6,8 +6,8 @@ import com.potato.balbambalbam.data.repository.*;
 import com.potato.balbambalbam.exception.InvalidUserNameException;
 import com.potato.balbambalbam.exception.SocialIdChangeException;
 import com.potato.balbambalbam.exception.UserNotFoundException;
-import com.potato.balbambalbam.user.join.jwt.dto.EditDto;
-import com.potato.balbambalbam.user.join.jwt.dto.JoinDto;
+import com.potato.balbambalbam.user.join.dto.EditDto;
+import com.potato.balbambalbam.user.join.dto.JoinDto;
 import com.potato.balbambalbam.user.join.jwt.JWTUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.transaction.Transactional;
